LayoutTests/platform/mac-monterey/imported/w3c/web-platform-tests/html/cross-origin-embedder-policy/anonymous-iframe/require-corp-embed-anonymous-iframe.tentative.https.window-expected.txt

CONSOLE MESSAGE: Refused to display 'https://localhost:9443/html/cross-origin-embedder-policy/resources/navigate-none.sub.html?postMessageTo=top' in a frame because of Cross-Origin-Embedder-Policy.

FAIL Loading an anonymous iframe with COEP: require-corp is allowed. assert_true: The anonymous iframe should be allowed. expected true got false
